Rename `BoundedContextName` to plural.
