+++ model/src/main/java/io/spine/examples/chatspn/package-info.java
@@ -27,7 +27,7 @@
 /**
  * This package contains basic data types of the ChatSPN application.
  */
-@BoundedContext(BoundedContextName.CHATS)
+@BoundedContext(CHATS)
 @CheckReturnValue
 @ParametersAreNonnullByDefault
 package io.spine.examples.chatspn;
@@ -36,3 +36,5 @@ import com.google.errorprone.annotations.CheckReturnValue;
 import io.spine.core.BoundedContext;
 
 import javax.annotation.ParametersAreNonnullByDefault;
+
+import static io.spine.examples.chatspn.BoundedContextNames.CHATS;
